pyproject.toml
[tool.poetry]
name = "py-sql-tasks"
version = "0.1.0"
description = ""
authors = ["Raphael Coelho <raphael.phael@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.1"
sqlalchemy = "^2.0.27"
tabulate = "^0.9.0"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
